backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
requests==2.31.0
psycopg2-binary==2.9.9
